backend_2/requirements.txt
Flask==2.3.2
PyJWT==2.8.0
filelock==3.12.2
Flask-Cors==4.0.0